requirements.txt
beautifulsoup4
requests
mysql-connector-python
sqlparse